# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  model.Permission:
    properties:
      access_level:
        type: string
      id:
        type: integer
      module:
        type: string
      role_id:
        type: integer
    type: object
...
  model.Role:
    properties:
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/model.Permission'
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Server for Ozinse Video Platform Backend
  title: Ozinse API
  version: "1.0"
paths:
...
  /api/roles:
    get:
      description: Returns a full list of roles, each containing its associated module
        permissions array. Requires administrator role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Role'
            type: array
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all roles (Admin only)
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Inserts a new role name and maps specified module permissions.
        Requires administrator role.
      parameters:
      - description: Role and Permissions Data
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/model.Role'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Role'
        "400":
          description: Invalid JSON input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new role with permissions (Admin only)
      tags:
      - roles
  /api/roles/{id}:
    delete:
      description: Deletes a role record by ID. Foreign key cascade constraints handle
        removing associated permissions automatically. Requires administrator role.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Role deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error or role not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a role (Admin only)
      tags:
      - roles
    get:
      description: Returns a single role object containing its mapped module permissions
        array. Requires administrator role.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Role'
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error or role not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a role by ID (Admin only)
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Modifies role details by ID and replaces old permissions with the
        new payload within a transaction. Requires administrator role.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Role Data
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/model.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Role'
        "400":
          description: Invalid ID or JSON input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error or role not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an existing role (Admin only)
      tags:
      - roles
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"